% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/getGridOffsetTimes.R
\name{getGridOffsetTimes}
\alias{getGridOffsetTimes}
\title{get time between trials by diffing Grid onset}
\usage{
getGridOffsetTimes(datafile = "alltimes_withExperiment.txt")
}
\arguments{
\item{datafile}{all-subject eprime timing/log as a tab delimited file}
}
\description{
this function generates a matrix of trial-to-trial gridonset to gridonset times.
each column is a run (1 to 4)
rows are trial num
N.B. expects to be run in same directory as alltimes_withExperiment.txt
    or given the file name
}
\examples{
ost<-getGridOffsetPerRun('path/to/textfile.txt')
 ost<-getGridOffsetPerRun() # if in same directory
}

